+++ packages/ssr_cli/lib/src/commands/create.dart
@@ -1,8 +1,6 @@
 import 'dart:io';
 import 'package:args/args.dart';
 import 'package:path/path.dart' as path;
-import '../utils/file_utils.dart';
-import '../utils/template_utils.dart';
 
 class CreateCommand {
   ArgParser get parser => ArgParser()