+++ packages/ssr_client/lib/src/navigation_service.dart
@@ -56,6 +56,31 @@ class NavigationService implements SsrRouter {
     _routeController.add(path);
   }
 
+  @override
+  bool matches(String pattern, String path) {
+    final regexPattern = pattern
+        .replaceAll(':', r'([^/]+)')
+        .replaceAll('/', r'\/');
+    final regex = RegExp('^$regexPattern\$');
+    return regex.hasMatch(path);
+  }
+
+  @override
+  Map<String, String> extractParams(String pattern, String path) {
+    final params = <String, String>{};
+    final patternParts = pattern.split('/');
+    final pathParts = path.split('/');
+    for (var i = 0; i < patternParts.length; i++) {
+      if (patternParts[i].startsWith(':')) {
+        final paramName = patternParts[i].substring(1);
+        if (i < pathParts.length) {
+          params[paramName] = pathParts[i];
+        }
+      }
+    }
+    return params;
+  }
+
   void dispose() {
     _routeController.close();
   }