+++ packages/ssr_router/lib/src/file_router.dart
@@ -42,7 +42,7 @@ class FileRouter {
     var routePath = '/' + pathParts
         .map((part) => part
             .replaceAll('_page.dart', '')
-            .replaceAll(RegExp(r'\[(\w+)\]'), ':$1'))
+            .replaceAll(RegExp(r'\[(\w+)\]'), ':\$1'))
         .join('/');
 
     // Handle index pages